05_hackathon/midterm_pipeline/api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
supabase>=2.3.0
httpx>=0.26.0
python-dotenv>=1.0.0
pydantic>=2.0.0
